copy over a few functions from _mixin_team.py to team.py
